+++ Attest/CountriesValidators/GreeceValidator.cs
@@ -18,7 +18,7 @@ namespace Attest.Countries
         /// <returns></returns>
         public override ValidationResult ValidateNationalIdentity(string number)
         {
-            number = number?.RemoveSpecialCharacthers();
+            number = number.RemoveSpecialCharacthers();
             if (!number.All(char.IsDigit))
             {
                 return ValidationResult.InvalidFormat("12345678901");